+++ packages/@expo/cli/src/start/server/webpack/resolveFromProject.ts
@@ -7,7 +7,10 @@ import { CommandError } from '../../../utils/errors';
 // TODO: Maybe combine with AsyncResolver?
 class WebpackImportError extends CommandError {
   constructor(projectRoot: string, moduleId: string) {
-    super('WEBPACK_IMPORT', `Missing package "${moduleId}" in the project at: ${projectRoot}`);
+    super(
+      'WEBPACK_IMPORT',
+      `Missing package "${moduleId}" in the project. Try running the command again. (cwd: ${projectRoot})`
+    );
   }
 }
 
@@ -25,7 +28,7 @@ function importFromProject(projectRoot: string, moduleId: string) {
 
 /** Import `webpack` from the project. */
 export function importWebpackFromProject(projectRoot: string): typeof import('webpack') {
-  return importFromProject(projectRoot, 'webpack').default;
+  return importFromProject(projectRoot, 'webpack');
 }
 
 /** Import `@expo/webpack-config` from the project. */
@@ -39,5 +42,5 @@ export function importExpoWebpackConfigFromProject(
 export function importWebpackDevServerFromProject(
   projectRoot: string
 ): typeof import('webpack-dev-server') {
-  return importFromProject(projectRoot, 'webpack-dev-server').default;
+  return importFromProject(projectRoot, 'webpack-dev-server');
 }